+++ packages/core/src/types/plugins.ts
@@ -1,6 +1,7 @@
+import type { ConfigRuleDefinitionObject } from "./configs.ts";
 import type { FilesValue } from "./files.ts";
-import type { AnyRule, Rule, RuleAbout } from "./rules.ts";
-import type { AnyOptionalSchema, InferredInputObject } from "./shapes.ts";
+import type { AnyRule, RuleAbout } from "./rules.ts";
+import type { InferredInputObject } from "./shapes.ts";
 
 /**
  * A Flint plugin containing a set of rules and presets.
@@ -20,7 +21,7 @@ export interface Plugin<
 		: Record<FilesKey & string, FilesValue>;
 
 	/**
-	 * The friendly name of the plugin, such as "JSON" or "Typescript".
+	 * The friendly name of the plugin, such as "JSON" or "TypeScript".
 	 */
 	name: string;
 
@@ -28,10 +29,7 @@ export interface Plugin<
 	 * Preset lists of rules to enable on files.
 	 * @see {@link https://flint.fyi/glossary#preset|flint.fyi/glossary#preset}
 	 */
-	presets: PluginPresets<
-		About,
-		NonNullable<Rules[number]["about"]["presets"]>[number]
-	>;
+	presets: PluginPresets<Rules>;
 
 	/**
 	 * Defines rules to configure or disable on files in a config.
@@ -44,13 +42,27 @@ export interface Plugin<
 	rulesById: Map<string, Rules[number]>;
 }
 
-export type PluginPresets<
-	About extends RuleAbout,
-	Presets extends string | undefined,
-> = Record<
-	Presets extends string ? Presets : never,
-	Rule<About, object, object, string, AnyOptionalSchema | undefined>[]
+export type PluginPresets<Rules extends AnyRule[]> = Record<
+	PluginPresetName<Rules>,
+	Rules[number][]
 >;
+
+type PluginConfiguredRule<Rule extends AnyRule> = ConfigRuleDefinitionObject & {
+	options: Rule["options"] extends undefined
+		? boolean
+		: boolean | InferredInputObject<Rule["options"]>;
+	rule: Rule;
+};
+
+type PluginConfiguredRules<Rules extends AnyRule[]> = {
+	[Rule in Rules[number] as Rule["about"]["id"]]: PluginConfiguredRule<Rule>;
+}[Rules[number]["about"]["id"]][];
+
+type PluginPresetName<Rules extends AnyRule[]> = Rules[number] extends infer R
+	? R extends { about: { presets: readonly (infer P extends string)[] } }
+		? P
+		: never
+	: never;
 
 /**
  * Defines rules to configure or disable on files in a config.
@@ -58,9 +70,9 @@ export type PluginPresets<
  */
 export type PluginRulesFactory<Rules extends AnyRule[]> = (
 	rulesOptions: PluginRulesOptions<Rules>,
-) => Rules;
+) => PluginConfiguredRules<Rules>;
 
-export type PluginRulesOptions<Rules extends AnyRule[]> = {
+type PluginRulesOptions<Rules extends AnyRule[]> = {
 	[Rule in Rules[number] as Rule["about"]["id"]]?: Rule["options"] extends undefined
 		? boolean
 		: boolean | InferredInputObject<Rule["options"]>;