+++ packages/site/src/components/getRuleForPlugin.ts
@@ -5,6 +5,7 @@ import { node } from "@flint.fyi/node";
 import { performance } from "@flint.fyi/performance";
 import { flint } from "@flint.fyi/plugin-flint";
 import { spelling } from "@flint.fyi/spelling";
+import { vitest } from "@flint.fyi/vitest";
 import { type AnyRule, json, md, packageJson, ts, yaml } from "flint";
 
 const plugins = {
@@ -19,6 +20,7 @@ const plugins = {
 	performance,
 	spelling,
 	ts,
+	vitest,
 	yaml,
 };
 
@@ -34,7 +36,7 @@ export function getRuleForPlugin(pluginId: string, ruleId: string): AnyRule {
 		throw new Error(`Unknown rule for ${pluginId}: ${ruleId}.`);
 	}
 
-	return rule as AnyRule;
+	return rule;
 }
 
 export function getRuleForPluginSafe(
@@ -52,5 +54,5 @@ export function getRuleForPluginSafe(
 		return undefined;
 	}
 
-	return rule as AnyRule;
+	return rule;
 }