+++ packages/ts/src/plugin.ts
@@ -248,6 +248,7 @@ import regexUnusedQuantifiers from "./rules/regexUnusedQuantifiers.ts";
 import regexValidity from "./rules/regexValidity.ts";
 import regexWordMatchers from "./rules/regexWordMatchers.ts";
 import regexZeroQuantifiers from "./rules/regexZeroQuantifiers.ts";
+import requireImports from "./rules/requireImports.ts";
 import responseJsonMethods from "./rules/responseJsonMethods.ts";
 import restrictedIdentifiers from "./rules/restrictedIdentifiers.ts";
 import returnAssignments from "./rules/returnAssignments.ts";
@@ -558,6 +559,8 @@ export const ts = createPlugin({
 		regexValidity,
 		regexWordMatchers,
 		regexZeroQuantifiers,
+		requireImports,
+		requireImports,
 		responseJsonMethods,
 		restrictedIdentifiers,
 		returnAssignments,